Move gating to content root
