+++ services/graphql-server/src/graphql/definitions/platform/content/interfaces/content.js
@@ -21,6 +21,7 @@ interface Content @requiresProject(fields: ["type"]) {
   # fields that used to be model specific, but were moved to the root
   deck: String @projection(localField: "mutations.Magazine.deck") @value(localField: "mutations.Magazine.deck")
   company(input: ContentCompanyInput = {}): ContentCompany @projection @refOne(loader: "platformContent", criteria: "contentCompany")
+  gating: ContentGating! @projection(localField: "mutations.Website.gating") @value(localField: "mutations.Website.gating")
 
   # fields from platform.trait::StatusEnabled
   status: Int @projection
