+++ services/graphql-server/src/graphql/definitions/platform/content/interfaces/index.js
@@ -3,7 +3,6 @@ const addressable = require('./addressable');
 const authorable = require('./authorable');
 const contactable = require('./contactable');
 const content = require('./content');
-const gateable = require('./gateable');
 const media = require('./media');
 
 module.exports = gql`
@@ -12,7 +11,6 @@ ${addressable}
 ${authorable}
 ${contactable}
 ${content}
-${gateable}
 ${media}
 
 `;
